requirements.txt
pdf2image>=1.0.0
Pillow>=8.0.0
pytesseract>=0.3.0
